feat: move Slack adapter from development to production
