+++ docs/pages/developers/development_platforms.md
@@ -1,12 +1,10 @@
 ## Development - Supported Clients
 
-In addition to our two clients approved for production use (Nextspace and Zoom), we also support other platforms:
+In addition to our three clients approved for production use (Nextspace, Zoom, and Slack), we also support other platforms:
 
 ❗ Caution! The platform types below are experimental! They are meant for evaluation, development and experimentation.
 ❗ Use them only with audiences who know they are engaging on an experimental platform!
 
 ✅ **[Nymspace](https://github.com/berkmancenter/internymspace_client)** - Visit the project for installation instructions
-
-✅ **Slack** (channels and DMs) - [installation guide](platforms/slack.md)
 
 ✅ **Your own!** Additional platforms can be supported by creating a new [adapter](adapters.md). Or, if you are developing a platform that can communicate directly with LLM Engine via our REST or Websocket APIs, there is no need to develop an adapter. Add your platform the to list of supported adapters in `adapters/config.ts`.
