+++ src/adapters/index.ts
@@ -1,12 +1,11 @@
 import zoom from './zoom.js'
-import slack from './development/slack/index.js'
+import slack from './slack/index.js'
 import config from '../config/config.js'
 
-const development = {
-  slack
-}
+const development = {}
 
 export default {
   ...(config.enableDevelopmentAdapters ? development : {}),
-  zoom
+  zoom,
+  slack
 }
