+++ tests/handlers/slack.handler.test.ts
@@ -1,15 +1,15 @@
 import httpStatus from 'http-status'
 import request from 'supertest'
 import crypto from 'crypto'
-import app from '../../../src/app.js'
-import setupIntTest from '../../utils/setupIntTest.js'
-import config from '../../../src/config/config.js'
-import { insertTopics } from '../../fixtures/topic.fixture.js'
-import Conversation from '../../../src/models/conversation.model.js'
-import { conversationAgentsEnabled, publicTopic } from '../../fixtures/conversation.fixture.js'
-import Adapter, { setAdapterTypes } from '../../../src/models/adapter.model.js'
-import webhookService from '../../../src/services/webhook.service.js'
-import defaultAdapterTypes from '../../../src/adapters/index.js'
+import app from '../../src/app.js'
+import setupIntTest from '../utils/setupIntTest.js'
+import config from '../../src/config/config.js'
+import { insertTopics } from '../fixtures/topic.fixture.js'
+import Conversation from '../../src/models/conversation.model.js'
+import { conversationAgentsEnabled, publicTopic } from '../fixtures/conversation.fixture.js'
+import Adapter, { setAdapterTypes } from '../../src/models/adapter.model.js'
+import webhookService from '../../src/services/webhook.service.js'
+import defaultAdapterTypes from '../../src/adapters/index.js'
 
 setupIntTest()
